Complete refactor of how optimized (combined) network filters are handled
Necessary to support the combination of lazy regex evaluation and binary deserialization. Previous implementation was running optimizations eagerly, before serialization, and storing computed regex in a new Network Filter. With such Network Filter was serialized and then deserialized again, RegexSet to match on could not be rebuilt correctly when requested.