Force unused load_network and load_cosmetic to true when serializing
Although these two fields are unused on new versions of this repo, they
are still serialized with dummy values to maintain backwards
compatibility.

In particular, the second field is checked by older versions of the repo
to determine if network blocking should be short-circuited. Thus, it's
important that this field stores a value of `true`.