rename native to js
`neon-cli` required binding code to be located in the `native`
directory, but now that we're not using it we're free to relocate it as
we please. I've wanted to do this for a while since `native` was a weird
name for JS bindings in a repo of primarily Rust code.