begin preparation for uBO's new templated scriptlet format
As of
https://github.com/gorhill/uBlock/commit/18a84d2819d49444fc31c5350677ecc5b2ec73c6,
the older template scriptlet parsing is no longer applicable. I'm
deprecating it here, and making it optional in the JS bindings. It's
still available for now if you need to continue using the older uBO
resources format, but going forwards, I recommend assembling
adblock-rust's `Resource` format directly using JS code.

If you need an example of this in practice, see
https://github.com/brave/brave-core-crx-packager/blob/master/lib/adBlockRustUtils.js