rename `debug-info` feature to `regex-debug-info`
Since no other debug info was provided, this is more descriptive. If
other types of debug info are desired in the future, additional features
and corresponding codepaths can be added.

This also means one single `RegexDebugInfo` struct can be used rather
than wrapping it in single-field structs twice while passing it up to
the `Engine`.