v0.9.5-brave-core - revert `idna` & `url` dependency updates
`brave-core` builds currently don't work well with the latest versions
of `url` and `idna`. This is due to idna's `"compiled_data"` feature
that is forcefully enabled through `url`'s dependencies.

This release temporarily downgrades those dependencies again, at least
until there can be a replacement for the `url` crate in `brave-core`
that uses Chromium's implementation of URL parsing. In the meantime, a
future change should replace usage of `idna::domain_to_ascii` with an
optional shim similar to how the `"embedded-domain-resolver"` feature
works.